Add files via upload
The neural network has three main layers. The input layer receives the 27 features from
the dataset. The hidden Layer consists of 128 neurons. This layer processes the inputs
from the previous layer, applying a non-linear transformation(ReLU) to capture complex
patterns in the data. The Output Layer, contains 7 neurons, corresponding to the 7 fault
categories in the dataset. This layer uses the softmax activation to produce probabilities
for each class. The network uses a fully connected architecture, where each neuron in one
layer is connected to every neuron in the subsequent layer. W1 and W2 represent the
weight matrices connecting the input layer to the hidden layer and the hidden layer to the
output layer, respectively.